[SL-7] menu component
