+++ src/utils/date.js
@@ -1,0 +1,9 @@
+const getDate = (dateObject) => {
+  return dateObject.toLocaleString('en-US', { day: '2-digit', month: '2-digit',year: 'numeric' });
+}
+
+const getDateIndex = (session, date) => {
+  return session.findIndex(storedDate =>  storedDate[0] === date);
+}
+
+export { getDate, getDateIndex };
